Inherit common websocket functionality from WebsocketBase
Move websocket code into websocket subdirectory
