+++ lib/ocpp1_6/websocket/CMakeLists.txt
@@ -1,0 +1,8 @@
+
+target_sources(ocpp
+    PRIVATE
+        websocket_base.cpp
+        websocket_plain.cpp
+        websocket_tls.cpp
+        websocket.cpp
+)
